skills/wechat-mp-illustrate/scripts/requirements.txt
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
markdown>=3.5.0
